# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Visitor registration, host approvals, agent tools, and dashboard data for checkIn.ai
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface OrganizationProfile {
  id: string;
  name: string;
  description: string;
  hostLabel: string;
  requiredFields: string[];
  notificationTarget: string;
  rulesSummary: string[];
}

export type VisitStatus = typeof VisitStatus[keyof typeof VisitStatus];


export const VisitStatus = {
  awaiting_approval: 'awaiting_approval',
  approved: 'approved',
  denied: 'denied',
  checked_in: 'checked_in',
  checked_out: 'checked_out',
} as const;

export type AgentTraceStepStatus = typeof AgentTraceStepStatus[keyof typeof AgentTraceStepStatus];


export const AgentTraceStepStatus = {
  completed: 'completed',
  waiting: 'waiting',
  blocked: 'blocked',
} as const;

export interface AgentTraceStep {
  agent: string;
  detail: string;
  status: AgentTraceStepStatus;
}

export interface VisitorVisit {
  id: string;
  profileId: string;
  visitorName: string;
  /** @nullable */
  affiliation?: string | null;
  hostName: string;
  purpose: string;
  /** @nullable */
  unitNumber?: string | null;
  /** @nullable */
  groupSize?: number | null;
  /** @nullable */
  visitSlot?: string | null;
  status: VisitStatus;
  createdBy: string;
  createdAt: string;
  /** @nullable */
  decisionAt?: string | null;
  /** @nullable */
  checkInAt?: string | null;
  /** @nullable */
  checkOutAt?: string | null;
  /** @nullable */
  decisionNote?: string | null;
}

export interface VisitorVisitInput {
  profileId: string;
  /** @minLength 1 */
  visitorName: string;
  affiliation?: string;
  /** @minLength 1 */
  hostName: string;
  /** @minLength 1 */
  purpose: string;
  unitNumber?: string;
  /** @minimum 1 */
  groupSize?: number;
  visitSlot?: string;
  /** @minLength 1 */
  createdBy: string;
}

export interface RegistrationDraftInput {
  profileId: string;
  /**
     * @minLength 3
     * @maxLength 2000
     */
  utterance: string;
}

export interface RegistrationDraft {
  profileId: string;
  /** @nullable */
  visitorName?: string | null;
  /** @nullable */
  affiliation?: string | null;
  /** @nullable */
  hostName?: string | null;
  /** @nullable */
  purpose?: string | null;
  /** @nullable */
  unitNumber?: string | null;
  /** @nullable */
  groupSize?: number | null;
  /** @nullable */
  visitSlot?: string | null;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  missingFields: string[];
  needsClarification: boolean;
  agentTrace: AgentTraceStep[];
}

export type HostDecisionInputAction = typeof HostDecisionInputAction[keyof typeof HostDecisionInputAction];


export const HostDecisionInputAction = {
  allow: 'allow',
  deny: 'deny',
} as const;

export interface HostDecisionInput {
  action: HostDecisionInputAction;
  /** @minLength 1 */
  actorName: string;
  note?: string;
}

export interface ActorActionInput {
  /** @minLength 1 */
  actorName: string;
}

export type VisitorNotificationStatus = typeof VisitorNotificationStatus[keyof typeof VisitorNotificationStatus];


export const VisitorNotificationStatus = {
  pending: 'pending',
  approved: 'approved',
  denied: 'denied',
} as const;

export interface VisitorNotification {
  id: string;
  profileId: string;
  visitId: string;
  recipient: string;
  title: string;
  message: string;
  status: VisitorNotificationStatus;
  createdAt: string;
}

export interface VisitorAuditEvent {
  id: string;
  profileId: string;
  /** @nullable */
  visitId?: string | null;
  actorName: string;
  eventType: string;
  message: string;
  occurredAt: string;
}

export type ActivityCellWeekday = typeof ActivityCellWeekday[keyof typeof ActivityCellWeekday];


export const ActivityCellWeekday = {
  Mon: 'Mon',
  Tue: 'Tue',
  Wed: 'Wed',
  Thu: 'Thu',
  Fri: 'Fri',
  Sat: 'Sat',
  Sun: 'Sun',
} as const;

export interface ActivityCell {
  weekday: ActivityCellWeekday;
  /**
     * @minimum 0
     * @maximum 23
     */
  hour: number;
  /** @minimum 0 */
  count: number;
}

export interface VisitorDashboard {
  totalVisits: number;
  currentlyInside: number;
  pendingApprovals: number;
  deniedToday: number;
  weeklyChangePercent: number;
  peakWindow: string;
  activity: ActivityCell[];
  recentVisits: VisitorVisit[];
}

export interface VisitorAnalyticsInput {
  profileId: string;
  /**
     * @minLength 3
     * @maxLength 500
     */
  question: string;
}

export interface VisitorAnalyticsAnswer {
  answer: string;
  metric: string;
  value: number;
  agentTrace: AgentTraceStep[];
}

export interface AudioTranscriptionInput {
  file: Blob;
}

export interface AudioTranscription {
  /** @nullable */
  requestId: string | null;
  transcript: string;
  /** @nullable */
  languageCode: string | null;
  /** @nullable */
  languageProbability: number | null;
}

export type GetVisitorDashboardParams = {
profileId: string;
dateRange?: GetVisitorDashboardDateRange;
};

export type GetVisitorDashboardDateRange = typeof GetVisitorDashboardDateRange[keyof typeof GetVisitorDashboardDateRange];


export const GetVisitorDashboardDateRange = {
  week: 'week',
  month: 'month',
} as const;

export type ListVisitorVisitsParams = {
profileId: string;
status?: VisitStatus;
search?: string;
};

export type ListVisitorNotificationsParams = {
profileId: string;
recipient: string;
};

export type ListVisitorAuditEventsParams = {
profileId: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};