# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Visitor registration, host approvals, agent tools, and dashboard data for checkIn.ai
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/registration-draft:
    post:
      operationId: createRegistrationDraft
      tags: [agents]
      summary: Extract a visitor registration draft from natural language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationDraftInput"
      responses:
        "200":
          description: Extracted draft and agent trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationDraft"
        "503":
          description: The AI provider is not configured or temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/analytics:
    post:
      operationId: askVisitorAnalytics
      tags: [agents]
      summary: Answer a natural-language question using approved read-only metrics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorAnalyticsInput"
      responses:
        "200":
          description: Read-only analytics answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorAnalyticsAnswer"
        "400":
          description: Unsupported analytics question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/transcribe:
    post:
      operationId: transcribeVisitorAudio
      tags: [agents]
      summary: Transcribe a short audio recording with Sarvam Saaras v3
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AudioTranscriptionInput"
      responses:
        "200":
          description: Audio transcription and detected language
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioTranscription"
        "400":
          description: Audio is missing, empty, or unsupported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Audio exceeds the upload size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: The speech-to-text provider is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AgentTraceStep:
      type: object
      properties:
        agent:
          type: string
        detail:
          type: string
        status:
          type: string
          enum: [completed, waiting, blocked]
      required: [agent, detail, status]
...
    RegistrationDraftInput:
      type: object
      properties:
        profileId:
          type: string
        utterance:
          type: string
          minLength: 3
          maxLength: 2000
      required: [profileId, utterance]
    RegistrationDraft:
      type: object
      properties:
        profileId:
          type: string
        visitorName:
          type: ["string", "null"]
        affiliation:
          type: ["string", "null"]
        hostName:
          type: ["string", "null"]
        purpose:
          type: ["string", "null"]
        unitNumber:
          type: ["string", "null"]
        groupSize:
          type: ["integer", "null"]
        visitSlot:
          type: ["string", "null"]
        confidence:
          type: number
          minimum: 0
          maximum: 1
        missingFields:
          type: array
          items:
            type: string
        needsClarification:
          type: boolean
        agentTrace:
          type: array
          items:
            $ref: "#/components/schemas/AgentTraceStep"
      required: [profileId, confidence, missingFields, needsClarification, agentTrace]
...
    VisitorAnalyticsInput:
      type: object
      properties:
        profileId:
          type: string
        question:
          type: string
          minLength: 3
          maxLength: 500
      required: [profileId, question]
    VisitorAnalyticsAnswer:
      type: object
      properties:
        answer:
          type: string
        metric:
          type: string
        value:
          type: integer
        agentTrace:
          type: array
          items:
            $ref: "#/components/schemas/AgentTraceStep"
      required: [answer, metric, value, agentTrace]
    AudioTranscriptionInput:
      type: object
      properties:
        file:
          type: string
          format: binary
      required: [file]
    AudioTranscription:
      type: object
      properties:
        requestId:
          type: ["string", "null"]
        transcript:
          type: string
        languageCode:
          type: ["string", "null"]
        languageProbability:
          type: ["number", "null"]
      required: [requestId, transcript, languageCode, languageProbability]