# lib/api-zod/src/generated/types/audioTranscriptionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Visitor registration, host approvals, agent tools, and dashboard data for checkIn.ai
 * OpenAPI spec version: 0.1.0
 */

export interface AudioTranscriptionInput {
  file: Blob;
}