# lib/api-zod/src/generated/types/registrationDraftInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Visitor registration, host approvals, agent tools, and dashboard data for checkIn.ai
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationDraftInput {
  profileId: string;
  /**
     * @minLength 3
     * @maxLength 2000
     */
  utterance: string;
}